update float representation

+++ core/src/test/java/com/airbnb/aerosolve/core/function/MultiDimensionPointTest.java
@@ -12,9 +12,9 @@ import static org.junit.Assert.*;
 public class MultiDimensionPointTest {
   @Test
   public void testEquals() throws Exception {
-    List<Float> a = Arrays.asList((float)3.0, (float)-5.0, (float)-9.0);
-    List<Float> a1 = Arrays.asList((float)3.0, (float)-5.0, (float)-9.0);
-    List<Float> b = Arrays.asList((float)3.0, (float)-5.0, (float)10.0);
+    List<Float> a = Arrays.asList(3.0f, -5.0f, -9.0f);
+    List<Float> a1 = Arrays.asList(3.0f, -5.0f, -9.0f);
+    List<Float> b = Arrays.asList(3.0f, -5.0f, 10.0f);
 
     MultiDimensionPoint p1 = new MultiDimensionPoint(a);
     MultiDimensionPoint p2 = new MultiDimensionPoint(a1);
@@ -30,14 +30,14 @@ public class MultiDimensionPointTest {
     List<Double> min = Arrays.asList(3.0, -5.0, -9.0);
     List<Double> max = Arrays.asList(13.0, 15.0, 10.0);
     List<List<Float>> r = Arrays.asList(
-        Arrays.asList((float)3.0, (float)-5.0, (float)-9.0),
-        Arrays.asList((float)13.0, (float)-5.0, (float)-9.0),
-        Arrays.asList((float)3.0, (float)15.0, (float)-9.0),
-        Arrays.asList((float)13.0, (float)15.0, (float)-9.0),
-        Arrays.asList((float)3.0, (float)-5.0, (float)10.0),
-        Arrays.asList((float)13.0, (float)-5.0, (float)10.0),
-        Arrays.asList((float)3.0, (float)15.0, (float)10.0),
-        Arrays.asList((float)13.0, (float)15.0, (float)10.0)
+        Arrays.asList(3.0f, -5.0f, -9.0f),
+        Arrays.asList(13.0f, -5.0f, -9.0f),
+        Arrays.asList(3.0f, 15.0f, -9.0f),
+        Arrays.asList(13.0f, 15.0f, -9.0f),
+        Arrays.asList(3.0f, -5.0f, 10.0f),
+        Arrays.asList(13.0f, -5.0f, 10.0f),
+        Arrays.asList(3.0f, 15.0f, 10.0f),
+        Arrays.asList(13.0f, 15.0f, 10.0f)
     );
 
     Map<List<Float>, MultiDimensionPoint> points = new HashMap<>();
@@ -68,10 +68,10 @@ public class MultiDimensionPointTest {
     List<List<Float>> keys = MultiDimensionPoint.getCombination(min, max);
     assertEquals(4, keys.size());
     List<List<Float>> r = Arrays.asList(
-        Arrays.asList((float)3.0, (float)5.0),
-        Arrays.asList((float)13.0, (float)5.0),
-        Arrays.asList((float)3.0, (float)15.0),
-        Arrays.asList((float)13.0, (float)15.0));
+        Arrays.asList(3.0f, 5.0f),
+        Arrays.asList(13.0f, 5.0f),
+        Arrays.asList(3.0f, 15.0f),
+        Arrays.asList(13.0f, 15.0f));
     assertEquals(keys, r);
 
     min = Arrays.asList(3.0, -5.0, -9.0);
@@ -79,37 +79,37 @@ public class MultiDimensionPointTest {
     keys = MultiDimensionPoint.getCombination(min, max);
     assertEquals(8, keys.size());
     r = Arrays.asList(
-        Arrays.asList((float)3.0, (float)-5.0, (float)-9.0),
-        Arrays.asList((float)13.0, (float)-5.0, (float)-9.0),
-        Arrays.asList((float)3.0, (float)15.0, (float)-9.0),
-        Arrays.asList((float)13.0, (float)15.0, (float)-9.0),
-        Arrays.asList((float)3.0, (float)-5.0, (float)10.0),
-        Arrays.asList((float)13.0, (float)-5.0, (float)10.0),
-        Arrays.asList((float)3.0, (float)15.0, (float)10.0),
-        Arrays.asList((float)13.0, (float)15.0, (float)10.0)
+        Arrays.asList(3.0f, -5.0f, -9.0f),
+        Arrays.asList(13.0f, -5.0f, -9.0f),
+        Arrays.asList(3.0f, 15.0f, -9.0f),
+        Arrays.asList(13.0f, 15.0f, -9.0f),
+        Arrays.asList(3.0f, -5.0f, 10.0f),
+        Arrays.asList(13.0f, -5.0f, 10.0f),
+        Arrays.asList(3.0f, 15.0f, 10.0f),
+        Arrays.asList(13.0f, 15.0f, 10.0f)
     );
     assertEquals(keys, r);
   }
 
   @Test
   public void testZero() {
-    float[] af = new float[]{0, 1, -2, (float) 3.4, (float)5.0, (float)-6.7,(float) 8.9};
-    List<Float> al = Arrays.asList((float)0, (float)1, (float)-2, (float) 3.4, (float)5.0, (float)-6.7,(float) 8.9);
-    List<Float> b = Arrays.asList((float)0, (float)1.0, (float) -2.0, (float)3.4, (float)5.0, (float)-6.7, (float)8.9);
-    assertEquals(0, MultiDimensionPoint.euclideanDistance(af, al), 0d);
-    assertEquals(0, MultiDimensionPoint.euclideanDistance(af, b), 0d);
+    float[] af = new float[]{0f, 1f, -2f,  3.4f, 5.0f, -6.7f, 8.9f};
+    List<Float> al = Arrays.asList(0f, 1f, -2f,  3.4f, 5.0f, -6.7f, 8.9f);
+    List<Float> b = Arrays.asList(0f, 1.0f,  -2.0f, 3.4f, 5.0f, -6.7f, 8.9f);
+
+    assertEquals(0, MultiDimensionPoint.euclideanDistance(af, al), 0);
+    assertEquals(0, MultiDimensionPoint.euclideanDistance(af, b), 0);
   }
 
   @Test
   public void test() {
-    float[] af = new float[]{(float)1.0, (float)-2.0, (float)3.0, (float)4.0};
-    float[] bf = new float[]{(float)-5.0, (float)-6.0, (float)7.0, (float)8.0};
-    List<Float> a = Arrays.asList((float)1.0, (float)-2.0, (float)3.0, (float)4.0);
-
-    List<Float> b = Arrays.asList((float)-5.0, (float)-6.0, (float)7.0, (float)8.0);
+    float[] af = new float[]{1.0f, -2.0f, 3.0f, 4.0f};
+    float[] bf = new float[]{-5.0f, -6.0f, 7.0f, 8.0f};
+    List<Float> a = Arrays.asList(1.0f, -2.0f, 3.0f, 4.0f);
+    List<Float> b = Arrays.asList(-5.0f, -6.0f, 7.0f, 8.0f);
 
     final double expected = Math.sqrt(84);
-    assertEquals(expected, MultiDimensionPoint.euclideanDistance(af, b), 0.001d);
-    assertEquals(expected, MultiDimensionPoint.euclideanDistance(bf, a), 0.001d);
+    assertEquals(expected, MultiDimensionPoint.euclideanDistance(af, b), 0.001);
+    assertEquals(expected, MultiDimensionPoint.euclideanDistance(bf, a), 0.001);
   }
 }
